refactor 시설예약 코드 리팩터링

+++ src/main/java/com/mycompany/project/reservation/query/dto/FacilityDTO.java
@@ -1,5 +1,6 @@
 package com.mycompany.project.reservation.query.dto;
 
+
 import lombok.*;
 
 @Getter
@@ -8,7 +9,7 @@ import lombok.*;
 @AllArgsConstructor
 @ToString
 public class FacilityDTO {
-    private int facilityId;
-    private String name;
-    private String status;
+  private Long facilityId;
+  private String name;
+  private String status;
 }
